Rename loading to io
